+++ src/chatbot/schedule.py
@@ -5,7 +5,7 @@ import random
 
 from apscheduler.schedulers.background import BackgroundScheduler
 
-import database_old as db
+import database as db
 
 from .events import generate_event
 from .model import Model
