+++ tests/test_chatbot/test_main.py
@@ -7,13 +7,13 @@ from datetime import timedelta
 
 import pytest
 
-import database_old as db
+import database as db
 from chatbot import Model
 from tests.test_chatbot.test_model import model
-from tests.test_database_old.test_characters import char_1
-from tests.test_database_old.test_main import db_init
-from tests.test_database_old.test_threads import thread_1
-from tests.test_database_old.test_users import user_1
+from tests.test_database.test_characters import char_1
+from tests.test_database.test_main import db_init
+from tests.test_database.test_threads import thread_1
+from tests.test_database.test_users import user_1
 
 main_module = importlib.import_module("chatbot.main")
 _get_system_message = getattr(main_module, "_get_system_message")
