+++ tests/test_chatbot/test_posts.py
@@ -10,16 +10,16 @@ from unittest.mock import MagicMock
 import pytest
 from google.cloud import storage
 
-import database_old as db
+import database as db
 from chatbot import Model, generate_social_media_post
 from tests.test_chatbot.test_model import model
-from tests.test_database_old.test_characters import char_1
-from tests.test_database_old.test_events import event_1, event_2
-from tests.test_database_old.test_main import db_init
-from tests.test_database_old.test_messages import message_1, message_2
-from tests.test_database_old.test_posts import post_1, post_2
-from tests.test_database_old.test_threads import thread_1
-from tests.test_database_old.test_users import user_1
+from tests.test_database.test_characters import char_1
+from tests.test_database.test_events import event_1, event_2
+from tests.test_database.test_main import db_init
+from tests.test_database.test_messages import message_1, message_2
+from tests.test_database.test_posts import post_1, post_2
+from tests.test_database.test_threads import thread_1
+from tests.test_database.test_users import user_1
 
 types_module = importlib.import_module("chatbot.types")
 ChatMessage = getattr(types_module, "ChatMessage")
