+++ tests/test_chatbot/test_response.py
@@ -5,14 +5,14 @@
 import importlib
 from datetime import datetime, timedelta, timezone
 
-import database_old as db
+import database as db
 from chatbot import Model, new_model, response_cycle
 from tests.test_chatbot.test_model import model
-from tests.test_database_old.test_characters import char_1
-from tests.test_database_old.test_main import db_init
-from tests.test_database_old.test_messages import message_1, message_2
-from tests.test_database_old.test_threads import thread_1
-from tests.test_database_old.test_users import user_1
+from tests.test_database.test_characters import char_1
+from tests.test_database.test_main import db_init
+from tests.test_database.test_messages import message_1, message_2
+from tests.test_database.test_threads import thread_1
+from tests.test_database.test_users import user_1
 
 types_module = importlib.import_module("chatbot.types")
 ChatMessage = getattr(types_module, "ChatMessage")
