+++ tests/test_routes/test_messages.py
@@ -10,17 +10,13 @@ import time
 import pytest
 from flask.testing import FlaskClient
 
-import database_old as db
+import database as db
 from tests.test_app import app, client
-from tests.test_database_old.test_characters import char_1
-from tests.test_database_old.test_main import db_init
-from tests.test_database_old.test_messages import (
-    message_1,
-    message_2,
-    scheduled_message,
-)
-from tests.test_database_old.test_threads import thread_1
-from tests.test_database_old.test_users import user_1
+from tests.test_database.test_characters import char_1
+from tests.test_database.test_main import db_init
+from tests.test_database.test_messages import message_1, message_2, scheduled_message
+from tests.test_database.test_threads import thread_1
+from tests.test_database.test_users import user_1
 
 
 def test_get_messages_by_thread(
