+++ tests/test_routes/test_threads.py
@@ -7,13 +7,13 @@ This file contains the tests for the routes/threads.py file.
 
 from flask.testing import FlaskClient
 
-import database_old as db
+import database as db
 from tests.test_app import app, client
-from tests.test_database_old.test_characters import char_1, char_2
-from tests.test_database_old.test_main import db_init
-from tests.test_database_old.test_messages import message_1
-from tests.test_database_old.test_threads import many_threads, thread_1
-from tests.test_database_old.test_users import user_1, user_2
+from tests.test_database.test_characters import char_1, char_2
+from tests.test_database.test_main import db_init
+from tests.test_database.test_messages import message_1
+from tests.test_database.test_threads import many_threads, thread_1
+from tests.test_database.test_users import user_1, user_2
 
 
 def test_new_thread(client: FlaskClient, user_1: db.User, char_1: db.Character) -> None:
